refactor: define arquivo unico para exportar os styles do auth
