+++ styles/auth/index.js
@@ -1,0 +1,4 @@
+export { createRegisterStyles } from './register.styles';
+export { styles as authInputStyles } from './auth-input.styles';
+export { styles as pawBackgroundStyles } from './paw-background.styles';
+export { styles as dogPeekingStyles } from './dog-peeking.styles';
